+++ lib/constants/onboarding_data.dart
@@ -3,8 +3,8 @@ List<Map<String, String>> onboardingData = [
     "image_url": "assets/health-check.png",
     "text": "To be a responsible donor, you must get a check-up",
   },
-  // {
-  //   "image_url": "assets/health-check.png",
-  //   "text": "To be a responsible donor, you must get a check-up",
-  // },
+  {
+    "image_url": "assets/health-check.png",
+    "text": "To be a responsible donor, you must get a check-up",
+  },
 ];
