+++ lib/main.dart
@@ -1,4 +1,4 @@
-import 'package:bloodbank_app/screens/onboarding1.dart';
+import 'package:bloodbank_app/screens/onboarding_screen.dart';
 import 'package:bloodbank_app/screens/splash_screen.dart';
 import 'package:flutter/material.dart';
 
@@ -14,7 +14,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: MySplashScreen(),
+      home: MyOnboardingPage(),
     );
   }
 }
